+++ packages/docusaurus-utils/src/__tests__/__fixtures__/simple-site/doc with space.md
@@ -1,0 +1,1 @@
+# Hoo hoo, if this path tricks you...
