+++ packages/docusaurus-utils/src/__tests__/__fixtures__/simple-site/hello.md
@@ -1,0 +1,7 @@
+---
+id: hello
+title: Hello, World !
+slug: /
+---
+
+Hello
