+++ packages/docusaurus-utils/src/__tests__/lastUpdateUtils.test.ts
@@ -1,0 +1,206 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import {jest} from '@jest/globals';
+import fs from 'fs-extra';
+import path from 'path';
+import {createTempRepo} from '@testing-utils/git';
+import shell from 'shelljs';
+import {getFileLastUpdate, readLastUpdateData} from '@docusaurus/utils';
+
+const convertDate = (date: number | undefined) =>
+  typeof date === 'number' ? date * 1000 : undefined;
+
+describe('getFileLastUpdate', () => {
+  const {repoDir} = createTempRepo();
+
+  const existingFilePath = path.join(
+    __dirname,
+    '__fixtures__/simple-site/hello.md',
+  );
+  it('existing test file in repository with Git timestamp', async () => {
+    const lastUpdateData = await getFileLastUpdate(existingFilePath);
+    expect(lastUpdateData).not.toBeNull();
+
+    const {author, timestamp} = lastUpdateData!;
+    expect(author).not.toBeNull();
+    expect(typeof author).toBe('string');
+
+    expect(timestamp).not.toBeNull();
+    expect(typeof timestamp).toBe('number');
+  });
+
+  it('existing test file with spaces in path', async () => {
+    const filePathWithSpace = path.join(
+      __dirname,
+      '__fixtures__/simple-site/doc with space.md',
+    );
+    const lastUpdateData = await getFileLastUpdate(filePathWithSpace);
+    expect(lastUpdateData).not.toBeNull();
+
+    const {author, timestamp} = lastUpdateData!;
+    expect(author).not.toBeNull();
+    expect(typeof author).toBe('string');
+
+    expect(timestamp).not.toBeNull();
+    expect(typeof timestamp).toBe('number');
+  });
+
+  it('non-existing file', async () => {
+    const consoleMock = jest
+      .spyOn(console, 'warn')
+      .mockImplementation(() => {});
+    const nonExistingFileName = '.nonExisting';
+    const nonExistingFilePath = path.join(
+      __dirname,
+      '__fixtures__',
+      nonExistingFileName,
+    );
+    await expect(getFileLastUpdate(nonExistingFilePath)).rejects.toThrow(
+      /An error occurred when trying to get the last update date/,
+    );
+    expect(consoleMock).toHaveBeenCalledTimes(0);
+    consoleMock.mockRestore();
+  });
+
+  it('git does not exist', async () => {
+    const mock = jest.spyOn(shell, 'which').mockImplementationOnce(() => null);
+    const consoleMock = jest
+      .spyOn(console, 'warn')
+      .mockImplementation(() => {});
+    const lastUpdateData = await getFileLastUpdate(existingFilePath);
+    expect(lastUpdateData).toBeNull();
+    expect(consoleMock).toHaveBeenLastCalledWith(
+      expect.stringMatching(
+        /.*\[WARNING\].* Sorry, the docs plugin last update options require Git\..*/,
+      ),
+    );
+
+    consoleMock.mockRestore();
+    mock.mockRestore();
+  });
+
+  it('temporary created file that is not tracked by git', async () => {
+    const consoleMock = jest
+      .spyOn(console, 'warn')
+      .mockImplementation(() => {});
+    const tempFilePath = path.join(repoDir, 'file.md');
+    await fs.writeFile(tempFilePath, 'Lorem ipsum :)');
+    await expect(getFileLastUpdate(tempFilePath)).resolves.toBeNull();
+    expect(consoleMock).toHaveBeenCalledTimes(1);
+    expect(consoleMock).toHaveBeenLastCalledWith(
+      expect.stringMatching(/not tracked by git./),
+    );
+    await fs.unlink(tempFilePath);
+  });
+});
+
+describe('readLastUpdateData', () => {
+  const defaultDevDate = 1539502055;
+  const defaultDevAuthor = 'Author';
+  const testDate = '2021-01-01';
+  const testDateTime = new Date(testDate).getTime();
+  const testAuthor = 'ozaki';
+
+  it('read last time show author time', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: true},
+      {date: testDate},
+    );
+    expect(convertDate(lastUpdatedAt)).toEqual(testDateTime);
+    expect(lastUpdatedBy).toBe(defaultDevAuthor);
+  });
+
+  it('read last author show author time', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: true},
+      {author: testAuthor},
+    );
+    expect(lastUpdatedBy).toEqual(testAuthor);
+    expect(lastUpdatedAt).toBe(defaultDevDate);
+  });
+
+  it('read last all show author time', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: true},
+      {author: testAuthor, date: testDate},
+    );
+    expect(lastUpdatedBy).toEqual(testAuthor);
+    expect(convertDate(lastUpdatedAt)).toEqual(testDateTime);
+  });
+
+  it('read last default show none', async () => {
+    const lastUpdate = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: false, showLastUpdateTime: false},
+      {},
+    );
+    expect(lastUpdate).toEqual({});
+  });
+
+  it('read last author show none', async () => {
+    const lastUpdate = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: false, showLastUpdateTime: false},
+      {author: testAuthor},
+    );
+    expect(lastUpdate).toEqual({});
+  });
+
+  it('read last time show author', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: false},
+      {date: testDate},
+    );
+    expect(lastUpdatedBy).toBe(defaultDevAuthor);
+    expect(lastUpdatedAt).toBeUndefined();
+  });
+
+  it('read last author show author', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: false},
+      {author: testAuthor},
+    );
+    expect(lastUpdatedBy).toBe('ozaki');
+    expect(lastUpdatedAt).toBeUndefined();
+  });
+
+  it('read last default show author default', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: true, showLastUpdateTime: false},
+      {},
+    );
+    expect(lastUpdatedBy).toBe(defaultDevAuthor);
+    expect(lastUpdatedAt).toBeUndefined();
+  });
+
+  it('read last time show time', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: false, showLastUpdateTime: true},
+      {date: testDate},
+    );
+    expect(lastUpdatedBy).toBeUndefined();
+    expect(convertDate(lastUpdatedAt)).toEqual(testDateTime);
+  });
+
+  it('read last author show time', async () => {
+    const {lastUpdatedAt, lastUpdatedBy} = await readLastUpdateData(
+      '',
+      {showLastUpdateAuthor: false, showLastUpdateTime: true},
+      {author: testAuthor},
+    );
+    expect(lastUpdatedBy).toBeUndefined();
+    expect(lastUpdatedAt).toEqual(defaultDevDate);
+  });
+});
